+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * POL_System_PartInfo identifies filesystems thru mount point instead of device
   (Btrfs subvolumes compatibility)
 * Modify bash/document_reader to pass extra arguments unchanged (http://www.playonlinux.com/en/topic-12519-Pass_arguments_to_PDFXChange_Viewer.html)
+* Remove "skipped lines" messages when the debugger gets the focus back
 
 # 4.2.5